requirements.txt
alembic==1.12.0
aniso8601==9.0.1
annotated-types==0.6.0
blinker==1.6.2
click==8.1.7
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
greenlet==3.0.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
Werkzeug==3.0.0
